docs(core): add developer doc comments and restructure tests into success/failure/edge dirs
Add detailed doc comments to all public/semi-public functions and types in
eval, coercion, context, functions, sum, if_fn, parser, and display modules.
Move inline #[cfg(test)] blocks from eval/mod.rs, coercion.rs, and context.rs
into per-function tests/{success,failure,edge}.rs directory structure, converting
coercion.rs and context.rs from flat files to mod.rs-based directories.
