+++ crates/core/src/eval/functions/logical/if_fn/mod.rs
@@ -4,6 +4,11 @@ use crate::eval::evaluate_expr;
 use crate::parser::ast::Expr;
 use crate::types::Value;
 
+/// `IF(condition, then, [else])` — conditional evaluation.
+///
+/// Accepts 2 or 3 arguments. The condition is evaluated and coerced to `bool`
+/// via [`to_bool`]. Only the taken branch is evaluated (short-circuit).
+/// With 2 arguments, the false case returns `Value::Bool(false)`.
 pub fn if_fn(args: &[Expr], ctx: &mut EvalCtx<'_>) -> Value {
     if let Some(err) = check_arity_len(args.len(), 2, 3) {
         return err;
