+++ crates/core/src/eval/functions/math/sum/mod.rs
@@ -2,6 +2,11 @@ use crate::eval::coercion::to_number;
 use crate::eval::functions::check_arity;
 use crate::types::{ErrorKind, Value};
 
+/// `SUM(n1, n2, ...)` — returns the arithmetic sum of 1–255 arguments.
+///
+/// Each argument is coerced to a number via [`to_number`]. The first error
+/// value encountered is returned immediately. Overflow to infinity returns
+/// `Value::Error(ErrorKind::Num)`.
 pub fn sum_fn(args: &[Value]) -> Value {
     if let Some(err) = check_arity(args, 1, 255) {
         return err;
