+++ crates/core/src/eval/mod.rs
@@ -11,8 +11,12 @@ use crate::types::{ErrorKind, Value};
 use coercion::{to_number, to_string_val};
 use functions::FunctionKind;
 
-/// Evaluate an expression tree, resolving variables from `ctx.ctx` and calling
-/// functions from `ctx.registry`.
+/// Walk an expression tree and produce a [`Value`].
+///
+/// Variables are resolved from `ctx.ctx`; functions are dispatched through
+/// `ctx.registry`. Eager functions receive pre-evaluated arguments; lazy
+/// functions (e.g. `IF`) receive raw [`Expr`] nodes and control their own
+/// evaluation order.
 pub fn evaluate_expr(expr: &Expr, ctx: &mut EvalCtx<'_>) -> Value {
     match expr {
         // ── Leaf nodes ──────────────────────────────────────────────────────
@@ -182,453 +186,4 @@ fn apply_cmp(op: &BinaryOp, ord: Option<std::cmp::Ordering>) -> bool {
 
 // ── Tests ────────────────────────────────────────────────────────────────────
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::eval::{Context, EvalCtx, Registry};
-    use crate::parser::ast::{BinaryOp, Expr, Span, UnaryOp};
-    use crate::types::{ErrorKind, Value};
-
-    fn dummy_span() -> Span { Span::new(0, 1) }
-
-    fn make_ctx() -> (Registry, Context) {
-        (Registry::new(), Context::empty())
-    }
-
-    fn run(expr: &Expr, registry: &Registry, ctx: Context) -> Value {
-        let mut eval_ctx = EvalCtx::new(ctx, registry);
-        evaluate_expr(expr, &mut eval_ctx)
-    }
-
-    // ── Leaf nodes ───────────────────────────────────────────────────────────
-
-    #[test]
-    fn leaf_number() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::Number(42.0, dummy_span());
-        assert_eq!(run(&expr, &reg, ctx), Value::Number(42.0));
-    }
-
-    #[test]
-    fn leaf_text() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::Text("hello".to_string(), dummy_span());
-        assert_eq!(run(&expr, &reg, ctx), Value::Text("hello".to_string()));
-    }
-
-    #[test]
-    fn leaf_bool_true() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::Bool(true, dummy_span());
-        assert_eq!(run(&expr, &reg, ctx), Value::Bool(true));
-    }
-
-    #[test]
-    fn leaf_bool_false() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::Bool(false, dummy_span());
-        assert_eq!(run(&expr, &reg, ctx), Value::Bool(false));
-    }
-
-    #[test]
-    fn variable_found() {
-        let (reg, _) = make_ctx();
-        let mut vars = std::collections::HashMap::new();
-        vars.insert("x".to_string(), Value::Number(7.0));
-        let ctx = Context::new(vars);
-        let expr = Expr::Variable("x".to_string(), dummy_span());
-        assert_eq!(run(&expr, &reg, ctx), Value::Number(7.0));
-    }
-
-    #[test]
-    fn variable_not_found() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::Variable("missing".to_string(), dummy_span());
-        assert_eq!(run(&expr, &reg, ctx), Value::Empty);
-    }
-
-    // ── Unary ops ────────────────────────────────────────────────────────────
-
-    #[test]
-    fn unary_neg_number() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::UnaryOp {
-            op: UnaryOp::Neg,
-            operand: Box::new(Expr::Number(5.0, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Number(-5.0));
-    }
-
-    #[test]
-    fn unary_neg_text_error() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::UnaryOp {
-            op: UnaryOp::Neg,
-            operand: Box::new(Expr::Text("abc".to_string(), dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Error(ErrorKind::Value));
-    }
-
-    #[test]
-    fn unary_percent() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::UnaryOp {
-            op: UnaryOp::Percent,
-            operand: Box::new(Expr::Number(50.0, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Number(0.5));
-    }
-
-    // ── Binary arithmetic ────────────────────────────────────────────────────
-
-    #[test]
-    fn binary_add() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Add,
-            left: Box::new(Expr::Number(3.0, dummy_span())),
-            right: Box::new(Expr::Number(4.0, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Number(7.0));
-    }
-
-    #[test]
-    fn binary_sub() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Sub,
-            left: Box::new(Expr::Number(10.0, dummy_span())),
-            right: Box::new(Expr::Number(3.0, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Number(7.0));
-    }
-
-    #[test]
-    fn binary_mul() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Mul,
-            left: Box::new(Expr::Number(6.0, dummy_span())),
-            right: Box::new(Expr::Number(7.0, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Number(42.0));
-    }
-
-    #[test]
-    fn binary_div() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Div,
-            left: Box::new(Expr::Number(10.0, dummy_span())),
-            right: Box::new(Expr::Number(4.0, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Number(2.5));
-    }
-
-    #[test]
-    fn binary_div_by_zero() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Div,
-            left: Box::new(Expr::Number(10.0, dummy_span())),
-            right: Box::new(Expr::Number(0.0, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Error(ErrorKind::DivByZero));
-    }
-
-    #[test]
-    fn binary_pow() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Pow,
-            left: Box::new(Expr::Number(2.0, dummy_span())),
-            right: Box::new(Expr::Number(10.0, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Number(1024.0));
-    }
-
-    // ── Concat ───────────────────────────────────────────────────────────────
-
-    #[test]
-    fn binary_concat() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Concat,
-            left: Box::new(Expr::Text("hello".to_string(), dummy_span())),
-            right: Box::new(Expr::Text(" world".to_string(), dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Text("hello world".to_string()));
-    }
-
-    // ── Comparisons ──────────────────────────────────────────────────────────
-
-    #[test]
-    fn cmp_eq_same_type_equal() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Eq,
-            left: Box::new(Expr::Number(5.0, dummy_span())),
-            right: Box::new(Expr::Number(5.0, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Bool(true));
-    }
-
-    #[test]
-    fn cmp_eq_same_type_unequal() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Eq,
-            left: Box::new(Expr::Number(5.0, dummy_span())),
-            right: Box::new(Expr::Number(6.0, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Bool(false));
-    }
-
-    #[test]
-    fn cmp_eq_cross_type() {
-        // Number vs Text: always false for Eq
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Eq,
-            left: Box::new(Expr::Number(1.0, dummy_span())),
-            right: Box::new(Expr::Text("1".to_string(), dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Bool(false));
-    }
-
-    #[test]
-    fn cmp_ne_cross_type() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Ne,
-            left: Box::new(Expr::Number(1.0, dummy_span())),
-            right: Box::new(Expr::Text("1".to_string(), dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Bool(true));
-    }
-
-    #[test]
-    fn cmp_lt_number_vs_text() {
-        // Number < Text in Excel ordering
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Lt,
-            left: Box::new(Expr::Number(999.0, dummy_span())),
-            right: Box::new(Expr::Text("a".to_string(), dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Bool(true));
-    }
-
-    #[test]
-    fn cmp_bool_ordering() {
-        // FALSE < TRUE
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Lt,
-            left: Box::new(Expr::Bool(false, dummy_span())),
-            right: Box::new(Expr::Bool(true, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Bool(true));
-    }
-
-    #[test]
-    fn cmp_text_gt_number() {
-        // Text > Number in Excel ordering
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Gt,
-            left: Box::new(Expr::Text("a".to_string(), dummy_span())),
-            right: Box::new(Expr::Number(999.0, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Bool(true));
-    }
-
-    // ── Function calls ───────────────────────────────────────────────────────
-
-    #[test]
-    fn function_unknown_returns_name_error() {
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::FunctionCall {
-            name: "NONEXISTENT".to_string(),
-            args: vec![],
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Error(ErrorKind::Name));
-    }
-
-    #[test]
-    fn function_eager_with_registered_fn() {
-        // Register a custom eager fn: DOUBLE(x) = x * 2
-        let mut reg = Registry::new();
-        reg.register_eager("DOUBLE", |args| {
-            if let Some(Value::Number(n)) = args.first() {
-                Value::Number(n * 2.0)
-            } else {
-                Value::Error(ErrorKind::Value)
-            }
-        });
-        let ctx = Context::empty();
-        let expr = Expr::FunctionCall {
-            name: "DOUBLE".to_string(),
-            args: vec![Expr::Number(21.0, dummy_span())],
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Number(42.0));
-    }
-
-    #[test]
-    fn function_eager_propagates_arg_error() {
-        let mut reg = Registry::new();
-        reg.register_eager("IDENTITY", |args| args.first().cloned().unwrap_or(Value::Empty));
-        let mut vars = std::collections::HashMap::new();
-        vars.insert("E".to_string(), Value::Error(ErrorKind::Ref));
-        let ctx = Context::new(vars);
-        let expr = Expr::FunctionCall {
-            name: "IDENTITY".to_string(),
-            args: vec![Expr::Variable("E".to_string(), dummy_span())],
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Error(ErrorKind::Ref));
-    }
-
-    // ── Error propagation in binary ops ──────────────────────────────────────
-
-    #[test]
-    fn error_propagates_left_in_add() {
-        let (reg, ctx) = make_ctx();
-        // "abc" + 1 => #VALUE! (left error from to_number)
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Add,
-            left: Box::new(Expr::Text("abc".to_string(), dummy_span())),
-            right: Box::new(Expr::Number(1.0, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Error(ErrorKind::Value));
-    }
-
-    #[test]
-    fn error_propagates_left_over_right_in_cmp() {
-        let (reg, _) = make_ctx();
-        let mut vars = std::collections::HashMap::new();
-        vars.insert("L".to_string(), Value::Error(ErrorKind::DivByZero));
-        vars.insert("R".to_string(), Value::Error(ErrorKind::Ref));
-        let ctx = Context::new(vars);
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Eq,
-            left: Box::new(Expr::Variable("L".to_string(), dummy_span())),
-            right: Box::new(Expr::Variable("R".to_string(), dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Error(ErrorKind::DivByZero));
-    }
-
-    // ── Non-finite arithmetic ────────────────────────────────────────────────
-
-    #[test]
-    fn pow_negative_fractional_exponent_is_num_error() {
-        // (-8.0).powf(0.333) → NaN → #NUM!
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Pow,
-            left: Box::new(Expr::Number(-8.0, dummy_span())),
-            right: Box::new(Expr::Number(0.333, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Error(ErrorKind::Num));
-    }
-
-    #[test]
-    fn pow_overflow_to_infinity_is_num_error() {
-        // 1e308.powf(2.0) → infinity → #NUM!
-        let (reg, ctx) = make_ctx();
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Pow,
-            left: Box::new(Expr::Number(1e308, dummy_span())),
-            right: Box::new(Expr::Number(2.0, dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Error(ErrorKind::Num));
-    }
-
-    // ── Lazy function dispatch ───────────────────────────────────────────────
-
-    #[test]
-    fn function_lazy_receives_raw_args() {
-        // Register a lazy fn that reads a variable from ctx — proves EvalCtx is passed correctly
-        // and args are NOT pre-evaluated.
-        let mut reg = Registry::new();
-        reg.register_lazy("LAZY_VAR", |args, ctx| {
-            // The first arg is a Variable expr; evaluate it manually to prove ctx is live
-            evaluate_expr(&args[0], ctx)
-        });
-        let mut vars = std::collections::HashMap::new();
-        vars.insert("X".to_string(), Value::Number(99.0));
-        let ctx = Context::new(vars);
-        let expr = Expr::FunctionCall {
-            name: "LAZY_VAR".to_string(),
-            args: vec![Expr::Variable("X".to_string(), dummy_span())],
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Number(99.0));
-    }
-
-    // ── Multi-arg eager error ordering ───────────────────────────────────────
-
-    #[test]
-    fn function_eager_first_error_wins_over_second() {
-        let mut reg = Registry::new();
-        reg.register_eager("ADD2", |args| {
-            match (args.get(0), args.get(1)) {
-                (Some(Value::Number(a)), Some(Value::Number(b))) => Value::Number(a + b),
-                _ => Value::Error(ErrorKind::Value),
-            }
-        });
-        // arg[0] = #DIV/0!, arg[1] = #REF! → first error should win
-        let mut vars = std::collections::HashMap::new();
-        vars.insert("A".to_string(), Value::Error(ErrorKind::DivByZero));
-        vars.insert("B".to_string(), Value::Error(ErrorKind::Ref));
-        let ctx = Context::new(vars);
-        let expr = Expr::FunctionCall {
-            name: "ADD2".to_string(),
-            args: vec![
-                Expr::Variable("A".to_string(), dummy_span()),
-                Expr::Variable("B".to_string(), dummy_span()),
-            ],
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Error(ErrorKind::DivByZero));
-    }
-
-    // ── Concat error propagation ─────────────────────────────────────────────
-
-    #[test]
-    fn concat_right_error_propagates() {
-        let (reg, _) = make_ctx();
-        let mut vars = std::collections::HashMap::new();
-        vars.insert("E".to_string(), Value::Error(ErrorKind::Value));
-        let ctx = Context::new(vars);
-        let expr = Expr::BinaryOp {
-            op: BinaryOp::Concat,
-            left: Box::new(Expr::Text("hello".to_string(), dummy_span())),
-            right: Box::new(Expr::Variable("E".to_string(), dummy_span())),
-            span: dummy_span(),
-        };
-        assert_eq!(run(&expr, &reg, ctx), Value::Error(ErrorKind::Value));
-    }
-}
+mod tests;
